% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvoTrees.R
\name{set_params}
\alias{set_params}
\title{Set EvoTree model parameters}
\usage{
set_params(loss = "linear", nrounds = 10L, lambda = 0, gamma = 0,
  eta = 0.1, max_depth = 5L, min_weight = 1, rowsample = 1,
  colsample = 1, nbins = 50, alpha = 0, metric = "none",
  seed = 444L, K = 1, ...)
}
\description{
Set EvoTree model parameters
}
